refactor: share renderPeerIssues between install error and peers check
