+++ deps/inspection/commands/tsconfig.json
@@ -55,6 +55,9 @@
       "path": "../../../installing/modules-yaml"
     },
     {
+      "path": "../../../installing/render-peer-issues"
+    },
+    {
       "path": "../../../lockfile/fs"
     },
     {
