+++ installing/render-peer-issues/test/index.ts
@@ -3,210 +3,112 @@ import { stripVTControlCharacters as stripAnsi } from 'node:util'
 import { expect, test } from '@jest/globals'
 import { renderPeerIssues } from '@pnpm/installing.render-peer-issues'
 
-test('renderPeerIssues()', () => {
-  expect(stripAnsi(renderPeerIssues({
-    'packages/0': {
-      conflicts: ['eee'],
-      intersections: { ddd: '^1.0.0' },
-      bad: {},
-      missing: {
-        ddd: [
-          {
-            parents: [
-              {
-                name: 'zzz',
-                version: '1.0.0',
-              },
-            ],
-            optional: false,
-            wantedRange: '^1.0.0',
-          },
-        ],
-        eee: [
-          {
-            parents: [
-              {
-                name: 'zzz',
-                version: '1.0.0',
-              },
-            ],
-            optional: false,
-            wantedRange: '^1.0.0',
-          },
-          {
-            parents: [
-              {
-                name: 'www',
-                version: '1.0.0',
-              },
-            ],
-            optional: false,
-            wantedRange: '^2.0.0',
-          },
-        ],
-      },
-    },
+test('renderPeerIssues() returns an empty string when there are no issues', () => {
+  expect(renderPeerIssues({
     '.': {
-      missing: {
-        aaa: [
-          {
-            parents: [
-              {
-                name: 'xxx',
-                version: '1.0.0',
-              },
-              {
-                name: 'yyy',
-                version: '1.0.0',
-              },
-            ],
-            optional: false,
-            wantedRange: '>=1.0.0 <3.0.0',
-          },
-        ],
-      },
-      bad: {
-        bbb: [
-          {
-            parents: [
-              {
-                name: 'xxx',
-                version: '1.0.0',
-              },
-            ],
-            foundVersion: '2',
-            resolvedFrom: [],
-            optional: false,
-            wantedRange: '^1.0.0',
-          },
-        ],
-        ccc: [
-          {
-            parents: [
-              {
-                name: 'xxx',
-                version: '1.0.0',
-              },
-              {
-                name: 'yyy',
-                version: '1.0.0',
-              },
-            ],
-            foundVersion: '2',
-            resolvedFrom: [
-              {
-                name: 'xxx',
-                version: '1.0.0',
-              },
-            ],
-            optional: false,
-            wantedRange: '^1.0.0',
-          },
-        ],
-      },
-      conflicts: [],
-      intersections: { aaa: '^1.0.0' },
-    },
-  }, { width: 500 }))).toMatchSnapshot()
-})
-
-test('renderPeerIssues() optional peer dependencies are printed only if they are in conflict with non-optional peers', () => {
-  expect(stripAnsi(renderPeerIssues({
-    '.': {
-      missing: {
-        aaa: [
-          {
-            parents: [
-              {
-                name: 'xxx',
-                version: '1.0.0',
-              },
-              {
-                name: 'yyy',
-                version: '1.0.0',
-              },
-            ],
-            optional: true,
-            wantedRange: '^1.0.0',
-          },
-          {
-            parents: [
-              {
-                name: 'xxx',
-                version: '1.0.0',
-              },
-              {
-                name: 'yyy',
-                version: '1.0.0',
-              },
-            ],
-            optional: false,
-            wantedRange: '^2.0.0',
-          },
-        ],
-        bbb: [
-          {
-            parents: [
-              {
-                name: 'xxx',
-                version: '1.0.0',
-              },
-            ],
-            optional: true,
-            wantedRange: '^1.0.0',
-          },
-        ],
-      },
-      bad: {},
-      conflicts: ['aaa'],
-      intersections: {},
-    },
-    empty: {
       missing: {},
       bad: {},
       conflicts: [],
       intersections: {},
     },
-  }, { width: 500 }))).toMatchSnapshot()
+  })).toBe('')
 })
 
-test('renderPeerIssues() format correctly the version ranges with spaces and "*"', () => {
+test('renderPeerIssues() renders bad peer dependencies', () => {
   expect(stripAnsi(renderPeerIssues({
     '.': {
+      missing: {},
+      bad: {
+        a: [
+          {
+            parents: [
+              { name: 'b', version: '1.0.0' },
+            ],
+            foundVersion: '2',
+            resolvedFrom: [],
+            optional: false,
+            wantedRange: '3',
+          },
+        ],
+      },
       conflicts: [],
-      intersections: { a: '*', b: '1 || 2' },
-      bad: {},
+      intersections: {},
+    },
+  }))).toMatchSnapshot()
+})
+
+test('renderPeerIssues() renders missing peer dependencies that are required', () => {
+  expect(stripAnsi(renderPeerIssues({
+    '.': {
       missing: {
         a: [
           {
             parents: [
-              {
-                name: 'z',
-                version: '1.0.0',
-              },
+              { name: 'b', version: '1.0.0' },
             ],
+            optional: false,
+            wantedRange: '^1.0.0',
+          },
+        ],
+      },
+      bad: {},
+      conflicts: [],
+      intersections: { a: '^1.0.0' },
+    },
+  }))).toMatchSnapshot()
+})
+
+test('renderPeerIssues() renders conflicting peer dependencies', () => {
+  expect(stripAnsi(renderPeerIssues({
+    '.': {
+      missing: {
+        a: [
+          {
+            parents: [{ name: 'b', version: '1.0.0' }],
+            optional: false,
+            wantedRange: '^1.0.0',
+          },
+          {
+            parents: [{ name: 'c', version: '1.0.0' }],
+            optional: false,
+            wantedRange: '^2.0.0',
+          },
+        ],
+      },
+      bad: {},
+      conflicts: ['a'],
+      intersections: {},
+    },
+  }))).toMatchSnapshot()
+})
+
+test('renderPeerIssues() formats version ranges with spaces or "*" with quotes', () => {
+  expect(stripAnsi(renderPeerIssues({
+    '.': {
+      missing: {
+        a: [
+          {
+            parents: [{ name: 'z', version: '1.0.0' }],
             optional: false,
             wantedRange: '*',
           },
         ],
         b: [
           {
-            parents: [
-              {
-                name: 'z',
-                version: '1.0.0',
-              },
-            ],
+            parents: [{ name: 'z', version: '1.0.0' }],
             optional: false,
             wantedRange: '1 || 2',
           },
         ],
       },
+      bad: {},
+      conflicts: [],
+      intersections: { a: '*', b: '1 || 2' },
     },
-  }, { width: 500 }))).toMatchSnapshot()
+  }))).toMatchSnapshot()
 })
 
-test('renderPeerIssues() do not fail if the parents array is empty', () => {
+test('renderPeerIssues() handles missing parents gracefully', () => {
   expect(stripAnsi(renderPeerIssues({
     '.': {
       missing: {
@@ -220,15 +122,7 @@ test('renderPeerIssues() do not fail if the parents array is empty', () => {
       },
       bad: {},
       conflicts: [],
-      intersections: {
-        foo: '^1.0.0',
-      },
+      intersections: { foo: '^1.0.0' },
     },
-  }, {
-    width: 500,
-  })).trim()).toBe(`.
-└─┬ <unknown> <unknown>
-  └── ✕ missing peer foo@">=1.0.0 <3.0.0"
-Peer dependencies that should be installed:
-  foo@^1.0.0`)
+  }))).toMatchSnapshot()
 })
