+++ pnpm-workspace.yaml
@@ -164,7 +164,6 @@ catalog:
   can-write-to-dir: ^2.0.0
   chalk: ^5.6.0
   ci-info: ^4.3.0
-  cli-columns: ^4.0.0
   cli-truncate: ^5.2.0
   cmd-extension: ^2.0.0
   comver-to-semver: ^2.0.0
